Added comments for various implementations added a fix for slow internet connection, the video was not getting played.
