+++ GSPlayer/Classes/Download/VideoDownloaderHandler.swift
@@ -170,27 +170,3 @@ private extension VideoDownloaderHandler {
         task?.resume()
     }
 }
-
-
-private extension UIDevice {
-    static var freeSpaceInBytes: Int64 {
-        if #available(iOS 11.0, *) {
-            if let space = try? URL(fileURLWithPath: NSHomeDirectory() as String).resourceValues(forKeys: [URLResourceKey.volumeAvailableCapacityForImportantUsageKey]).volumeAvailableCapacityForImportantUsage {
-                return space
-            } else {
-                return 0
-            }
-        } else {
-            if let systemAttributes = try? FileManager.default.attributesOfFileSystem(forPath: NSHomeDirectory() as String),
-                let freeSpace = (systemAttributes[FileAttributeKey.systemFreeSize] as? NSNumber)?.int64Value {
-                return freeSpace
-            } else {
-                return 0
-            }
-        }
-    }
-    
-    static var freeDiskSpaceInMB: Int64 {
-        return UIDevice.freeSpaceInBytes / (1024 * 1024)
-    }
-}
